Ask before reinstalling

+++ setup.sh
@@ -113,15 +113,16 @@ git submodule update --quiet --init --recursive
 print_result $? "Submodules fetched"
 
 # Brew stuff
-if [ "$(uname)" = "Darwin"]; then
+if [ "$(uname)" = "Darwin" ]; then
 if ! type "brew" >/dev/null; then
   print_error "No homebrew found"
   exit 1
 fi
 brew tap Homebrew/bundle
-print_info "Reinstalling packages"
-execute "brew bundle --file=packages/Brewfile" "Homebrew & Cask & Mac AppStore"
-execute "pip3 install -U -r packages/requirements3.txt" "pip3"
+if ask_for_confirmation "?Install packages?"; then
+  execute "brew bundle --file=packages/Brewfile" "Homebrew & Cask & Mac AppStore"
+  execute "pip3 install -U -r packages/requirements3.txt" "pip3"
+fi
 fi
 
 # Actual symlink stuff
